+++ website/pages/docs/_meta.ts
@@ -1,0 +1,25 @@
+export default {
+  "-Getting Started": {
+    title: "Getting Started",
+    type: "separator",
+  },
+  installation: {
+    title: "Installation",
+  },
+  configuration: {
+    title: "Configuration",
+  },
+  faq: {
+    title: "FAQ",
+  },
+  "-Other": {
+    title: "Other",
+    type: "separator",
+  },
+  packages: {
+    title: "Packages",
+  },
+  glossary: {
+    title: "Glossary",
+  },
+} as const;