+++ website/pages/docs/packages/_meta.ts
@@ -1,0 +1,7 @@
+export default {
+  "@eslint-react/eslint-plugin": {
+    title: "@eslint-react/eslint-plugin",
+    href: "https://www.npmjs.com/package/@eslint-react/eslint-plugin",
+    newWindow: true,
+  },
+} as const;